# rest/go/docs/swagger.yaml
basePath: /
definitions:
  task.Task:
    properties:
      deadline:
        description: Tags     []string   `json:"tags,omitempty"`
        format: date-time
        type: string
      id:
        type: string
      text:
        type: string
    type: object
info:
  contact: {}
  description: Task APIs
  title: Task APIs
  version: "1.0"
paths:
  /task:
    get:
      consumes:
      - application/json
      description: list all tasks
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/task.Task'
            type: array
        "400":
          description: error
          schema:
            type: string
        "404":
          description: error
          schema:
            type: string
        "500":
          description: error
          schema:
            type: string
      summary: List tasks
      tags:
      - task
    post:
      consumes:
      - application/json
      description: create task
      parameters:
      - description: Create task
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/task.Task'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: error
          schema:
            type: string
        "404":
          description: error
          schema:
            type: string
        "500":
          description: error
          schema:
            type: string
      summary: Create task
      tags:
      - task
  /task/{id}:
    delete:
      consumes:
      - application/json
      description: delete task by ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "404":
          description: error
          schema:
            type: string
      summary: Delete a task
      tags:
      - task
    get:
      consumes:
      - application/json
      description: get task by ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/task.Task'
        "404":
          description: error
          schema:
            type: string
      summary: Show a task
      tags:
      - task
    put:
      consumes:
      - application/json
      description: update task
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/task.Task'
            type: array
        "400":
          description: error
          schema:
            type: string
        "404":
          description: error
          schema:
            type: string
        "500":
          description: error
          schema:
            type: string
      summary: Update task
      tags:
      - task
swagger: "2.0"